pyproject.toml
[project]
name = "currency-exchange-api"
version = "0.1.0"
description = ""
authors = [
    {name = "Den_k0",email = "denys.kotovskiy@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "django (>=5.1.7,<6.0.0)",
    "djangorestframework (>=3.15.2,<4.0.0)",
    "djangorestframework-simplejwt (>=5.5.0,<6.0.0)",
    "psycopg2-binary (>=2.9.10,<3.0.0)",
    "python-dotenv (>=1.1.0,<2.0.0)",
    "ruff (>=0.11.2,<0.12.0)",
    "requests (>=2.32.3,<3.0.0)",
    "django-filter (>=25.1,<26.0)",
    "drf-spectacular (>=0.28.0,<0.29.0)"
]


[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"


[tool.ruff]
line-length = 79
exclude = [
    "currency/migrations/*",
    "users/migrations/*"
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "N",  # PEP8 naming convetions
]

ignore = ["I001", "E731", "F403", "F405"]

[tool.ruff.lint.flake8-quotes]
inline-quotes = "double"

[tool.ruff.lint.pydocstyle]
convention = "google"
